Handle 'if' change to '_types.Script' (#16)
In ProcessorBase, the 'if' type changed from string to _types.Script
which is a struct. The elasticsearch-specification allows 'if' to also
be a string through a special 'shortcutProperty' where when 'if' is
defined as a scalar then this means to populate the _types.Script#source
field.

Rather than drastically change the generated Go types and json-schema
this basically overrides the _types.Script as a string. In the future we
could consider properly handing the 'shortcutProperty'.

Closes #15